requirements.txt
markdown-it-py>=3.0.0
semchunk>=2.0.0
transformers>=4.21.0
tiktoken>=0.5.0
pydantic>=2.0.0